small cleanup and add json for William
