+++ dri/requirements.txt
@@ -1,3 +1,4 @@
 PyGithub>=1.43.8
+jsonpickle>=1.2
 colorama>=0.4.1
 six>=1.12.0
